+++ packages/adminui-vue/CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - `USelectCollection` prop `entityName` removed in flavor of computed property `entityName`
 
 ### Fixed
- - show actual text instead of empty <<< >>> in error window in case unhandled rejection message contains <<<text>>>    
+ - show actual text instead of empty <<< >>> in error window in case unhandled rejection message contains <<<text>>>
+ - prevent multiple emitting of `input` event in case UInput is of type `number`
 
 ## [1.9.10] - 2019-12-20
 ## [1.9.9] - 2019-12-19