+++ packages/adminui-vue/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - show actual text instead of empty <<< >>> in error window in case unhandled rejection message contains <<<text>>>
  - prevent multiple emitting of `input` event in case UInput is of type `number`
+ - prevent showing of form validation error twice in case user close tab on the unsaved form and validation error occures 
 
 ## [1.9.10] - 2019-12-20
 ## [1.9.9] - 2019-12-19